Update index.html and style.css
Uploaded images
Added facts and titles to index.html.
Added sources and links for the images.
Added styling to column background that corresponds with color of each planet.
Changed the font of the table headers and data.
Changed the font size column headers to be larger depending on the diameter of the planet.
Changed the row labels to be bold